Move cell view model into a separate file.
